clear code, add type module
