+++ index.html
@@ -93,13 +93,7 @@
       <button class="button button_type_delete"></button>
     </li>
   </template>
-  <script src="./scripts/elements.js"></script>
-  <script src="./scripts/FormValidator.js"></script>
-  <script src="./scripts/index.js"></script>
-  <script src="./scripts/Card.js"></script>
-
-  <!-- <script src="./scripts/validate.js"></script> -->
-  <!-- <script src="./scripts/valid-test.js"></script> -->
+  <script type="module" src="./scripts/index.js"></script>
 </body>
 
 </html>
